+++ philosophers/philo/includes/structures.h
@@ -6,7 +6,7 @@
 /*   By: dahkang <dahkang@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/28 20:32:38 by dahkang           #+#    #+#             */
-/*   Updated: 2023/02/03 01:40:35 by dahkang          ###   ########.fr       */
+/*   Updated: 2023/02/03 14:06:07 by dahkang          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,7 @@
 
 # include <pthread.h>
 # include <unistd.h>
+# include <stdint.h>
 
 # define TIME 0
 # define ORIGIN 1
@@ -41,11 +42,11 @@ enum	e_bool
 // unit of time is microsec
 typedef struct s_rules
 {
-	int	n_philo;
-	int	time_die;		
-	int	time_eat;		
-	int	time_sleep;		
-	int	n_must_eat;
+	int			n_philo;
+	uint64_t	time_die;		
+	uint64_t	time_eat;		
+	uint64_t	time_sleep;		
+	int			n_must_eat;
 }	t_rules;
 
 typedef struct s_args
@@ -56,7 +57,7 @@ typedef struct s_args
 	pthread_mutex_t		msg_lock;
 	pthread_mutex_t		common;
 	t_bool				all_done;
-	int					start_time;
+	uint64_t			start_time;
 }	t_args;
 
 typedef struct s_philo
@@ -67,7 +68,7 @@ typedef struct s_philo
 	pthread_mutex_t	*rfork;
 	pthread_t		tid;
 	int				eat_cnt;
-	int				last_eat_time;
+	uint64_t		last_eat_time;
 	int				done;
 }	t_philo;
